Adding footer to panels
I want to put a button in the footer to re-roll the style for a given
day. The first step is adding a footer. Making it hidden and then it'll
appear on rollover. I really wanted to do this in CSS but I'm having a
hell of a time overriding the .panel-footer class.